Move component away from page root
